requirements.txt
streamlit
speechrecognition
pyttsx3
google-generativeai
python-dotenv
comtypes
